feat: index.js: Convert **4. Some text** into \s2, improve Unicode
ChatGPT tends to fake headings with bold text, instead of using proper ## Markdown. So I convert it to Heading 2. Doesn't always correspond to the document hierarchy, but OpenAI are being inconsistent from one reply to the next.

Unicode conversion RTF_special_chars_escape() escapes the Private Use Area (PUA), a weird ChatGPT inclusion in the Markdown, which shows as unrecognized something in RTF.